fix: edit wager is now working
