+++ src/app/lib/data.ts
@@ -293,7 +293,7 @@ export const getOneWager = async (id: string) => {
 
 // edit wager
 export const editWagerWithId = async (id: string, body: any) => {
-    const res = await fetch(`/api/wagers/edit?wager_id=${id}`, {
+    const res = await fetch(`/api/wagers?wager_id=${id}`, {
         method: "PUT",
         body: JSON.stringify(body),
     });
